--- work1/ytxz999/Campus_IF_Love/Campus_IF_Love_Story.md
@@ -1,215 +1,0 @@
-# 玩法介绍
-
-游戏开始 -> 三位角色依次登场，分别带有两个选项，选中正确选项则进入该角色故事线（接下来的行为都以攻略该角色为主）否则进行下一位角色的选择 -> 选中主角后，开始增加好感度的各种尝试（对话/送礼物）-> 好感度满之后，解锁结局
-
-
-# 主角配置
-
-男主身份：某大学 2025 级新生
-
-特点：颜值和智商双双天花板，Python 编程大佬，喜欢扮猪吃老虎
-
-背景：开学典礼上作为新生代表发言，一举成为全场焦点
-
-性格设定：外表谦逊内心自信，不张扬，但遇到编程相关话题变得激动，进入癫狂状态
-
-目标：在大学生活中与不同角色邂逅，建立羁绊，解锁结局
-- -------------------
-
-2. 前置剧情（开局）
-
-（黑屏，逐字显示）
-“你是一名刚刚踏入大学校园的新生。”
-“在开学典礼上，拿下压倒性成绩第一的你被选为新生代表发言。”
-“在全场上千人的注视下，你气质非凡，发言流畅，很快成为焦点人物。”
-“消息迅速传开，关于‘神秘新生代表’的讨论充斥着整个校园。”
-“于是，在这个新的舞台上，你与三位不同的女生产生了交集……”
-
-3. 女主角介绍
-（1）黑长直学姐 —— 社团部长
-
-姓名：未定（可由玩家命名 or 默认“学姐”）
-
-身份：绘画社团的部长（大二）
-
-外貌：黑长直发，清冷气质，身材高挑
-
-性格：外冷内热，对新人要求严格，但心底善良，喜欢偷偷依赖男主
-
-相遇场景：
-玩家加入社团时遇见学姐，她对男主很感兴趣，通过一系列小举动靠近男主
-
-好感度关键点：
-
-- 聊天时如果展现出技术实力，好感度上升快
-- 展现出艺术能力和技巧，好感度上升快
-- 如果送"画笔"或"鲜花"，效果显著
-
-结局：
-与学姐一同升学考研，有情人终成眷属
-——————————————————————————————————————————————————
-对话库
-学姐：「你是新来的吧？画画过吗？」
-
-选项A：「其实我更擅长写代码，但我愿意试试画画。」
-
-选项B：「画画？那种小孩子的东西我可没兴趣。」
-
-学姐：「这幅画……构图还算不错，你是随便画的吗？」
-
-选项A：「嗯，我是随手画的，没想到能得到你的认可。」
-
-选项B：「不，我可是很认真的！因为你在看嘛。」
-
-学姐：「加入社团后，不能三天打鱼两天晒网。」
-
-选项A：「放心吧，我会认真对待。」
-
-选项B：「社团只是玩玩而已吧？不要太严肃了。」
-
-学姐：「我平时比较严格，你会不会觉得我很不好相处？」
-
-选项A：「严格才好啊，我喜欢有目标的人。」
-
-选项B：「嗯……确实有点难相处。」
-
-学姐：「今天的社团活动结束了，你要不要和我一起收拾？」
-
-选项A：「当然可以，我来帮你。」
-
-选项B：「算了吧，我还有事要做。」
-____________________________________________________
-
-
-（2）白毛萝莉 —— 图书馆少女
-
-姓名：未定（可由玩家命名 or 默认“小白”）
-
-身份：同级新生，爱泡图书馆
-
-外貌：白色及肩短发，身材娇小，穿着校服时喜欢戴大号耳机
-
-性格：有点天然呆，喜欢学习，对编程很感兴趣，像一只猫儿
-
-相遇场景：
-主角在图书馆借书时，发现她一个人抱着厚厚的《Python入门》
-
-好感度关键点：
-
-聊天时如果耐心讲解编程知识，好感度快速上升
-
-如果送“奶茶”或“编程笔记”，好感度暴涨
-
-如果送“奇怪的石头”，好感度下降（觉得主角恶趣味）
-
-结局：
-在图书馆的学习后，她对主角心生依赖，两人约定共同进步
-
-____________________________________
-对话库
-
-小白：「Python 里的 for 循环，我总是写不对……」
-
-选项A：「来，我教你一个小技巧！」
-
-选项B：「这种简单的东西，你都不会？」
-
-小白：「你看我借了好多书，会不会显得很傻？」
-
-选项A：「不会啊，这说明你很爱学习，很可爱。」
-
-选项B：「嗯……的确有点太贪心了。」
-
-小白：「写代码的时候，我总喜欢喝奶茶……你呢？」
-
-选项A：「我也是！来，下次我请你一杯。」
-
-选项B：「我只喝水，健康第一。」
-
-小白：「你会不会觉得我太依赖你了？」
-
-选项A：「依赖我也没关系，我喜欢这种感觉。」
-
-选项B：「嗯……是有点吧。」
-
-小白：「要不要一起留在图书馆自习？我想有人陪。」
-
-选项A：「好啊，我正好也要复习。」
-
-选项B：「算了，我还是回宿舍打游戏吧。」
-________________________________________________
-（3）知性姐姐 —— 校外咖啡店常客
-
-姓名：未定（默认“姐姐”）
-
-身份：研究生，偶尔在校外咖啡店看书
-
-外貌：成熟知性，披肩长发，喜欢穿风衣
-
-性格：温柔理性，话语间透露出阅历和智慧
-
-相遇场景：
-主角想在校外咖啡店寻求清净，偶遇学姐，并交谈甚欢
-
-好感度关键点：
-
-聊天时如果展现成熟稳重的一面，好感度提升明显
-
-她喜欢“鲜花”和“奶茶”，不喜欢“奇怪的石头”
-
-结局：
-在一次深夜的咖啡店长谈后，两人互诉心意，最终牵手走向未来
-————————————————————————————————————————————
-姐姐：「你也喜欢在咖啡店看书吗？」
-
-选项A：「是啊，这里很安静，很适合思考。」
-
-选项B：「其实我只是随便找个地方坐。」
-
-姐姐：「研究生的生活，其实没有你想象的那么光鲜。」
-
-选项A：「我能理解，压力一定很大吧。」
-
-选项B：「那我还是不要考研了。」
-
-姐姐：「你觉得，什么样的人最值得依靠？」
-
-选项A：「稳重冷静的人。」
-
-选项B：「长得好看的人。」
-
-姐姐：「我常常一个人坐到很晚，你不会觉得孤单吗？」
-
-选项A：「有时候孤单是种享受。」
-
-选项B：「一个人太寂寞了，我受不了。」
-
-姐姐：「如果有一天，我遇到困难，你会帮我吗？」
-
-选项A：「当然会，你不用一个人扛着。」
-
-选项B：「我大概帮不了你吧。」
-——————————————————————————————————————————————
-4. 系统元素关联（方便写代码）
-
-道具（示例）
-
-鲜花（+10 对大部分角色）
-
-编程笔记（+15 姐姐/小白）
-
-奶茶（+20 学姐/小白/姐姐）
-
-奇怪的石头（-10 所有人）
-
-精致的钢笔（+20 姐姐/学姐 +10 小白）
-
-可爱玩偶（学姐/姐姐+10 小白+20）
-
-夜宵外卖：小白（+5），姐姐（-5），学姐（+0）
-
-可以自己再增加！
-
-好感度阈值：100 解锁结局
-
-互动方式：聊天（+5 基础值），送礼物（不同效果）
