--- work1/ytxz999/Campus_IF_Love/README.md
@@ -1,47 +1,0 @@
-# 校园·if 恋（Campus IF Love）
-
-## 友情提示
-
-这是一个以攻略为核心的恋爱模拟游戏（Galgame），目标是完成对三位女主角的攻略。为了熟悉剧情，你应该阅读文件夹中的[剧本](Campus_IF_Love_Story.md)。
-
-## 游戏玩法
-
-1. 完成场景的介绍。
-2. 你将面临第一个决定点，选择是否进入第一位女主角的专属路线。
-3. 若你在上一步失败或放弃，将面临第二个决定点，选择是否进入第二位女主角的路线。
-4. 若前两次机会都错过了，你将迎来最后的决定点。若再次失败，游戏将直接结束。一旦成功选定一位女主角，后续的决定点将自动跳过。
-
-## 项目结构与文档要求
-
-### 你要完成的函数及内容
-
-1. 角色类中的**对话函数** `talk()`，你可以选择通过 `get` 方法直接获得字典中的对话内容，也可以直接使用 `if` 语句根据不同角色设定不同的对话。
-   - **提示**：对于同一个角色，你可以根据好感的不同或者对话次数的不同设置不同的对话以及选项。
-
-2. 角色类中的**送礼函数** `give_gift()`，你可以直接通过 `get` 方法获得好感度字典中的元素，也可以直接使用 `if in` 等语句判断不同的礼物对于不同的角色的好感度加成。
-   - **提示**：通过调用 `change_affinity(value)` 函数，并传入相应的好感度 value 值。
-
-3. 第二位女主角的**决定**选择题（可以借鉴第一位）。
-
-4. 第三位女主角的**决定**选择题（可以借鉴第一位）。
-
-5. 主线剧情中玩家的**操作函数**，包括：选择 1 进行对话时的逻辑，选择 2 进行送礼时的逻辑。
-
-### 你仍需要完成
-
-1. 在作业完成之后，写一份文档讲述一下你遇到了什么问题，如何解决的（如果没有遇到问题就不写）。
-2. 你需要将你的项目代码合理地拆分到至少三个不同的 `.py` 文件中（如 `game.py`、`manage.py`、`story.py`）。
-3. 使用 Type Hint 类型注释。
-
-### Bonus
-
-1. 完成存档、读档、即时存档功能。
-   - **提示**：可以为剧情编写索引，先将剧情分为几块，然后为每一句进行编码。然后记录当前的剧情块编码和剧情句编码，并且记录当前的所有数值，写入 JSON 串。读档的时候定位到剧情块编码和剧情句编码，用存档中的数值暴力覆盖即可。即时存档的意思是专门设计一个存储空间，每次操作都进行存档。
-
-2. 自由拓展。
-
-## 注意事项
-
-1. 答案仅供参考，且不涉及「你仍需要完成」和「Bonus」的部分，请勿直接抄袭。
-2. 相关的对话库与礼物好感度数据，已作为字典提供在 `.py` 文件的底部。
-3. 如果有需要，随意修改框架代码或者字典里面的内容，只需要创造出你最喜欢的游戏即可。
